Winter, Day 0 [Tealc#6853]
Starting out in Evac Shelter C-41 our heroine bears witness to the tragic death of our new friend. After scouting the surrounding area and cleaning the two nearby evac shelters of useful foods she sets out for the refugee center to the south-west. Braving packs of feral dogs, coyotes, and even zombified bobcats she eventually made it to the refugee center only to be turned away. She talks to those in the lobby and learns of their situation, giving a few of them food to help their aching bellies.