Winter, Day 1 [Scrivener07#4154]
I took a moment to organize my belongings before.
The Refugee Center isnt much but its shelter with lights and some protection.
The heroine wonders if this could be a new home.

If this was to be home then one ought to find a way to be useful.
While getting to know some of the survivors in the shelter a Madelaine Escobar offered a job to help Shelter C-41.
Says they work for an "Old Guard", or whats left of the government.
Madelaine gave the task to kill a pair of bandits preying on locals. This kind of work suits our militia background and so agreed to look into this matter.

Some time will need to be taken for preperations. A scouting operation will be apporiate to start with. Madaline reports the bandits to be about 106tls away, to far for today.
For now I should see if there is any food or water in the shelter perimeter. There is some water but not much ready to eat.

I got back to the lobby after search around inside the perimeter fence and a loud piercing shrill was heard from the back of the building.
It was a big panic. Everyone with a weapon ran to the source of the screams. The living dead were everywhere but the refugees took them down quickly.
I got covered is guts. Looks like these guys can handling themselves. Unfortunately Rhyzaea Johnny Died in the incident.