Day 3 - Exploration
08:00
Started out in the ranch kitchen, holding a bunch of junk, a m1 garand
with 10 shots, and a m1911 with 25.
Dumped out a bunch of the junk and moved items around from the kitchen
into the living room area.

From there, went outside and herded the cows into one of the buildings,
using the rope from the haybale building (smash the bales) to tie them
down.

Then moved the cow corpse up front to next to the crates, and
dismembered the zombie cow corpses, sticking the useful fat also close
to the crates.

09:00
With domestic affairs wrapped up I hopped onto the bike outside, and
went to go check out the world around us.

I first headed down south, where I discovered an LMOE shelter. It had
either already been looted or was poorly stocked, but I grabbed
everything of value from it.
Notable loot I took includes a ski jacket and a 6 round shotgun.
The items of note that I left are: a 5 round shotgun, a blunderbuss, and
an electric jackhammer.

10:00
With all the loot from the LMOE shelter either loaded into the bike or
my jacket, I set off east towards the refugee center and houses around
there.

I discovered a dump just south of the refugee center is filled with
bandits. I briefly considered taking it on a bit, but I probably had
about a single bullet between my rifle and my pistol for bandit there...
Let's skip that.
I did steal a jumper cable from the office on the outskirts, though.

Past that, I continued on east, where I found a quite intact Mechanized
Infantry Carrier. I grabbed a 9mm pistol from it, left some .50 BMG ammo
in it, marked it on the map, and moved on, following the road as it
curved south.

The cave on that road has a cougar hiding in it, if we find ourselves
needing meat in the future.

11:00
Continuing down the road, and skipping yet another small town, I found a
radio tower. Climbing to the top of this radio tower, I discovered a map
and a survivor telescope, giving me a great view of the map.

I continued on down the road a little bit, and found a pristine solar
car, only in need of hotwiring.
While attempting to do this, I was ambushed by a mi-go scout and a
normal mi-go. Luckily, I was holding my m1 garand, and after 3 shots,
both were heavily bleeding, and fled. I suspect they died, but am not
sure about that.

I then spent most of the rest of the day trying, and failing, to start
the car.

19:00
It started getting dark, so I gave up, and looted the nearby cabin (only
really got a wood axe), grabbed the first quest from Hub 01, then
discovered the other dump was also full of NPC bandits.

I couldn't find the robot to kill, so I just grabbed the EMP grenades
and headed back to base.

When I got back to base, I moved items around more and set up a little
workshop area with a our loot sorted out.
Midway through that, I got down enough water to deal with the mild
hypovolemia from the earlier mi-go attack.

03:00
When the organizing was all wrapped up, I went to sleep.

13:00
When I woke up, I organized our inventory a bit, then retired my turn.